Removed utils
