+++ src/event/source/windows.rs
@@ -2,13 +2,13 @@ use std::time::Duration;
 
 use crossterm_winapi::{Console, Handle, InputEventType, KeyEventRecord, MouseEvent};
 
-use crate::event::{sys::windows::WinApiPoll, Event};
+use crate::event::{sys::windows::poll::WinApiPoll, Event};
 
 #[cfg(feature = "event-stream")]
 use super::super::sys::Waker;
 use super::super::{
     source::EventSource,
-    sys::windows::{handle_key_event, handle_mouse_event},
+    sys::windows::parse::{handle_key_event, handle_mouse_event},
     timeout::PollTimeout,
     InternalEvent, Result,
 };
