--- src/utils/sys.rs
@@ -1,5 +1,0 @@
-#[cfg(windows)]
-pub(crate) mod windows;
-
-#[cfg(unix)]
-pub(crate) mod unix;
